artf3271 fix
git-svn-id: http://code.jaspersoft.com/svn/repos/jasperreports/trunk/jasperreports@2221 76e25fc1-081a-0410-bb1b-c96bc70638b7

+++ src/net/sf/jasperreports/engine/export/JRTextExporter.java
@@ -570,7 +570,7 @@ public class JRTextExporter extends JRAbstractExporter
 	 */
 	protected int calculateYCoord(int y)
 	{
-		return pageHeight * y / jasperPrint.getPageHeight();
+		return (int) (((long) pageHeight * y) / jasperPrint.getPageHeight());
 	}
 
 	/**
